fix: fixes install script after tests

+++ install.sh
@@ -2,9 +2,11 @@
 
 INSTALL_CRON=false
 DOC=/usr/local/share/applications/migration_factory
-CONFIG=/etc/migration_factory
+CONFIG_DESTINATION_PATH=/etc/migration_factory
 ENDPOINT_FILE=endpoints.json
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
+
+set -e
 
 display_help() {
   echo "$0"
@@ -16,10 +18,11 @@ display_help() {
 }
 
 check_arguments() {
-  while test $# -gt 0; do
+  while test "$#" -gt 0; do
     case "$1" in
     -h | --help)
       display_help
+      exit 0
       ;;
     --cron)
       INSTALL_CRON=true
@@ -33,25 +36,29 @@ check_arguments() {
   done
 }
 
-sudo cp scripts/* /usr/local/bin/
-sudo mkdir -p $CONFIG
+check_arguments "$@"
 
-if [ -f "$CONFIG/$ENDPOINT_FILE.new" ]; then
-    cmp --silent $ENDPOINT_FILE "$CONFIG/$ENDPOINT_FILE.new" || cp $ENDPOINT_FILE "$CONFIG/$ENDPOINT_FILE.new"
+sudo cp scripts/* /usr/local/bin/
+sudo mkdir -p "$CONFIG_DESTINATION_PATH"
+
+if [ -f "$CONFIG_DESTINATION_PATH/$ENDPOINT_FILE.new" ]; then
+    cmp --silent "config/$ENDPOINT_FILE" "$CONFIG_DESTINATION_PATH/$ENDPOINT_FILE.new" || cp "config/$ENDPOINT_FILE" "$CONFIG_DESTINATION_PATH/$ENDPOINT_FILE.new"
 else
-  if [ -f "$CONFIG/$ENDPOINT_FILE" ]; then
-    cmp --silent $ENDPOINT_FILE "$CONFIG/$ENDPOINT_FILE" || cp $ENDPOINT_FILE "$CONFIG/$ENDPOINT_FILE.new"
+  if [ -f "$CONFIG_DESTINATION_PATH/$ENDPOINT_FILE" ]; then
+    cmp --silent "config/$ENDPOINT_FILE" "$CONFIG_DESTINATION_PATH/$ENDPOINT_FILE" || cp "config/$ENDPOINT_FILE" "$CONFIG_DESTINATION_PATH/$ENDPOINT_FILE.new"
   else
-    cp $ENDPOINT_FILE "$CONFIG/$ENDPOINT_FILE"
+    cp "config/$ENDPOINT_FILE" "$CONFIG_DESTINATION_FILE/$ENDPOINT_FILE"
   fi
 fi
 
-sudo mkdir -p $DOC
-sudo cp config/0-import-tags.csv config/0-Migration-intake-form.csv $DOC
+sudo mkdir -p "$DOC"
+sudo cp config/0-import-tags.csv config/0-Migration-intake-form.csv "$DOC"
 
 if [ "$INSTALL_CRON" = true ] ; then
-    crontab -l > install_mf_script
-    echo "*/10 * * * * $DIR/$0" >> install_mf_script
-    crontab install_mf_script
-    rm install_mf_script
+    if [ `crontab -l | grep -q "*/10 * * * * $DIR/$0"` ] ; then
+        crontab -l > install_mf_script
+        echo "*/10 * * * * $DIR/$0" >> install_mf_script
+        crontab install_mf_script
+        rm install_mf_script
+    fi
 fi
